+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- `renderIfoFile` function.
 
 ## [0.1.0.1] - 2017-03-14
 ### Changed